+++ tests/testthat/test_LearnerFTTransformer.R
@@ -25,6 +25,45 @@ make_ft_transformer = function(task_type, ...) {
   invoke(lrn, .key = sprintf("%s.ft_transformer", task_type), .args = params)
 }
 
+no_wd = function(name) {
+  linear_bias_param = grepl("linear_", name, fixed = TRUE) && grepl(".bias", name, fixed = TRUE)
+
+  other_no_wd_params = c("embedding", "_normalization")
+
+  return(
+    any(map_lgl(other_no_wd_params, function(pattern) grepl(pattern, name, fixed = TRUE)))
+    || linear_bias_param
+  )
+}
+
+rtdl_param_groups = function(parameters) {
+  no_wd_idx = map_lgl(names(parameters), no_wd)
+  no_wd_group = parameters[no_wd_idx]
+
+  main_group = parameters[!no_wd_idx]
+
+  list(
+    list(params = main_group),
+    list(params = no_wd_group, weight_decay = 0)
+  )
+}
+
+test_that("param groups work", {
+  learner = make_ft_transformer("classif")
+  default_weight_decay = 0.23
+  learner$param_set$set_values(opt.weight_decay = default_weight_decay)
+  learner$param_set$set_values(opt.param_groups = rtdl_param_groups)
+
+  task = tsk("german_credit")$filter(1:10)
+  learner$train(task)
+
+  expect_equal(length(learner$model$optimizer$param_groups), 2L)
+  expect_equal(learner$model$optimizer$param_groups[[1L]]$weight_decay, default_weight_decay)
+  expect_equal(learner$model$optimizer$param_groups[[2L]]$weight_decay, 0)
+
+  expect_learner(learner)
+})
+
 test_that("basic functionality", {
   learner = make_ft_transformer("classif")
   task = tsk("german_credit")$filter(1:10)